tests/submission/Google/video/loop/loop_reflects_true.html

<!doctype html>
<html>
 <head>
  <title>video.loop - reflection test</title>
  <script src="../../../../resources/testharness.js"></script>
  <script src="../../../../resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-loop">spec reference</a></p>
  <video id="video" loop>
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_true(
  document.getElementById("video").loop,
  "videoElement.loop reflects true when present");
});
  </script>
 </body>
</html>
